+++ src/main/java/org/realityforge/bazel/depgen/config/JavaConfig.java
@@ -6,6 +6,8 @@ public final class JavaConfig
 {
   @Nullable
   private Boolean exportDeps;
+  @Nullable
+  private String alias;
 
   @Nullable
   public Boolean getExportDeps()
@@ -16,5 +18,16 @@ public final class JavaConfig
   public void setExportDeps( @Nullable final Boolean exportDeps )
   {
     this.exportDeps = exportDeps;
+  }
+
+  @Nullable
+  public String getAlias()
+  {
+    return alias;
+  }
+
+  public void setAlias( @Nullable final String alias )
+  {
+    this.alias = alias;
   }
 }